# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sources": {
      "post": {
        "description": "Register a new data source for metadata ingestion (WordPress, Slack, Sitemap, S3-Compatible Storage, Azure Blob Storage, Google Cloud Storage, PostgreSQL, MySQL, MSSQL, Oracle, Hive, Databricks, Snowflake, MongoDB, PowerBI, Tableau, Confluence, Jira, Service Desk).",
        "operationId": "SourcesController_createSource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceDto"
              },
              "examples": {
                "wordpress": {
                  "summary": "WordPress Source",
                  "value": {
                    "type": "WORDPRESS",
                    "name": "Production WordPress",
                    "config": {
                      "type": "WORDPRESS",
                      "required": {
                        "url": "https://blog.example.com"
                      },
                      "masked": {
                        "username": "admin",
                        "application_password": "your-application-password"
                      },
                      "optional": {
                        "content": {
                          "fetch_posts": true,
                          "fetch_pages": true
                        }
                      },
                      "sampling": {
                        "strategy": "RANDOM",
                        "limit": 25
                      }
                    }
                  }
                },
                "slack": {
                  "summary": "Slack Source",
                  "value": {
                    "type": "SLACK",
                    "name": "Production Slack",
                    "config": {
                      "type": "SLACK",
                      "required": {
                        "workspace": "acme"
                      },
                      "masked": {
                        "bot_token": "xoxb-your-bot-token"
                      },
                      "optional": {
                        "channels": {
                          "channel_types": [
                            "public_channel"
                          ]
                        },
                        "ingestion": {
                          "limit_total_messages": 5000
                        }
                      },
                      "sampling": {
                        "strategy": "RANDOM",
                        "limit": 100
                      }
                    }
                  }
                },
                "sitemap": {
                  "summary": "Sitemap Source",
                  "value": {
                    "type": "SITEMAP",
                    "name": "Production Website Index",
                    "config": {
                      "type": "SITEMAP",
                      "required": {
                        "sitemap_url": "https://www.example.com/sitemap.xml"
                      },
                      "optional": {
                        "crawl": {
                          "max_nested_sitemaps": 100
                        }
                      },
                      "sampling": {
                        "strategy": "RANDOM",
                        "limit": 25
                      }
                    }
                  }
                },
                "s3CompatibleStorage": {
                  "summary": "S3-Compatible Storage Source",
                  "value": {
                    "type": "S3_COMPATIBLE_STORAGE",
                    "name": "S3-compatible exports",
                    "config": {
                      "type": "S3_COMPATIBLE_STORAGE",
                      "required": {
                        "bucket": "customer-exports"
                      },
                      "masked": {
                        "aws_access_key_id": "access-key",
                        "aws_secret_access_key": "secret-key"
                      },
                      "optional": {
                        "connection": {
                          "endpoint_url": "https://storage.example.internal"
                        },
                        "scope": {
                          "prefix": "daily/",
                          "include_extensions": [
                            ".csv",
                            ".pdf"
                          ]
                        }
                      },
                      "sampling": {
                        "strategy": "LATEST",
                        "limit": 100
                      }
                    }
                  }
                },
                "azureBlobStorage": {
                  "summary": "Azure Blob Storage Source",
                  "value": {
                    "type": "AZURE_BLOB_STORAGE",
                    "name": "Azure blob exports",
                    "config": {
                      "type": "AZURE_BLOB_STORAGE",
                      "required": {
                        "account_url": "https://acme.blob.core.windows.net",
                        "container": "customer-exports"
                      },
                      "masked": {
                        "azure_account_key": "account-key"
                      },
                      "optional": {
                        "scope": {
                          "prefix": "daily/",
                          "include_extensions": [
                            ".csv",
                            ".pdf"
                          ]
                        }
                      },
                      "sampling": {
                        "strategy": "LATEST",
                        "limit": 100
                      }
                    }
                  }
                },
                "googleCloudStorage": {
                  "summary": "Google Cloud Storage Source",
                  "value": {
                    "type": "GOOGLE_CLOUD_STORAGE",
                    "name": "GCS exports",
                    "config": {
                      "type": "GOOGLE_CLOUD_STORAGE",
                      "required": {
                        "bucket": "customer-exports"
                      },
                      "masked": {
                        "gcp_credentials_json": "{\"type\":\"service_account\",\"project_id\":\"acme\"}"
                      },
                      "optional": {
                        "connection": {
                          "project_id": "acme-prod"
                        },
                        "scope": {
                          "prefix": "daily/",
                          "include_extensions": [
                            ".csv",
                            ".pdf"
                          ]
                        }
                      },
                      "sampling": {
                        "strategy": "LATEST",
                        "limit": 100
                      }
                    }
                  }
                },
                "postgresql": {
                  "summary": "PostgreSQL Source",
                  "value": {
                    "type": "POSTGRESQL",
                    "name": "Production PostgreSQL",
                    "config": {
                      "type": "POSTGRESQL",
                      "required": {
                        "host": "localhost",
                        "port": 5432
                      },
                      "masked": {
                        "username": "postgres",
                        "password": "test"
                      },
                      "optional": {
                        "scope": {
                          "database": "postgres"
                        }
                      },
                      "sampling": {
                        "strategy": "RANDOM",
                        "limit": 20
                      }
                    }
                  }
                },
                "mysql": {
                  "summary": "MySQL Source",
                  "value": {
                    "type": "MYSQL",
                    "name": "Production MySQL",
                    "config": {
                      "type": "MYSQL",
                      "required": {
                        "host": "localhost",
                        "port": 3306
                      },
                      "masked": {
                        "username": "root",
                        "password": "example"
                      },
                      "optional": {
                        "scope": {
                          "database": "app_db"
                        }
                      },
                      "sampling": {
                        "strategy": "RANDOM",
                        "limit": 20
                      }
                    }
                  }
                },
                "confluence": {
                  "summary": "Confluence Source",
                  "value": {
                    "type": "CONFLUENCE",
                    "name": "Engineering Confluence",
                    "config": {
                      "type": "CONFLUENCE",
                      "required": {
                        "base_url": "https://your-domain.atlassian.net",
                        "account_email": "user@example.com"
                      },
                      "masked": {
                        "api_token": "atlassian-api-token"
                      },
                      "optional": {
                        "scope": {
                          "spaces": {
                            "keys": [
                              "ENG"
                            ],
                            "status": "current"
                          }
                        },
                        "content": {
                          "include_attachments": true,
                          "include_footer_comments": true,
                          "include_inline_comments": true
                        }
                      },
                      "sampling": {
                        "strategy": "LATEST",
                        "limit": 50
                      }
                    }
                  }
                },
                "jira": {
                  "summary": "Jira Source",
                  "value": {
                    "type": "JIRA",
                    "name": "Platform Jira",
                    "config": {
                      "type": "JIRA",
                      "required": {
                        "base_url": "https://your-domain.atlassian.net",
                        "account_email": "user@example.com"
                      },
                      "masked": {
                        "api_token": "atlassian-api-token"
                      },
                      "optional": {
                        "scope": {
                          "project_keys": [
                            "PLAT"
                          ],
                          "jql": "statusCategory != Done"
                        },
                        "content": {
                          "include_comments": true,
                          "include_attachments": true
                        }
                      },
                      "sampling": {
                        "strategy": "LATEST",
                        "limit": 100
                      }
                    }
                  }
                },
                "servicedesk": {
                  "summary": "Service Desk Source",
                  "value": {
                    "type": "SERVICEDESK",
                    "name": "Support Service Desk",
                    "config": {
                      "type": "SERVICEDESK",
                      "required": {
                        "base_url": "https://your-domain.atlassian.net",
                        "account_email": "user@example.com"
                      },
                      "masked": {
                        "api_token": "atlassian-api-token"
                      },
                      "optional": {
                        "scope": {
                          "service_desk_ids": [
                            1
                          ],
                          "request_status": "OPEN_REQUESTS"
                        },
                        "content": {
                          "include_comments": true,
                          "include_attachments": true
                        }
                      },
                      "sampling": {
                        "strategy": "LATEST",
                        "limit": 100
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Source successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - validation failed"
          }
        },
        "summary": "Create a new data source",
        "tags": [
          "Sources"
        ]
      },
      "get": {
        "description": "Retrieve a list of all registered data sources",
        "operationId": "SourcesController_listSources",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of sources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all data sources",
        "tags": [
          "Sources"
        ]
      }
    },
    "/sources/{id}": {
      "get": {
        "description": "Retrieve detailed information about a specific data source",
        "operationId": "SourcesController_getSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Source unique identifier",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Source details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          }
        },
        "summary": "Get source by ID",
        "tags": [
          "Sources"
        ]
      },
      "put": {
        "description": "Update the configuration and/or name of an existing data source",
        "operationId": "SourcesController_updateSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Source unique identifier",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceDto"
              },
              "examples": {
                "updateConfig": {
                  "summary": "Update source configuration",
                  "value": {
                    "config": {
                      "type": "WORDPRESS",
                      "required": {
                        "url": "https://updated-blog.example.com"
                      },
                      "masked": {
                        "username": "admin",
                        "application_password": "updated-application-password"
                      },
                      "optional": {
                        "content": {
                          "fetch_posts": true,
                          "fetch_pages": false
                        }
                      },
                      "sampling": {
                        "strategy": "RANDOM",
                        "limit": 25
                      }
                    }
                  }
                },
                "updateName": {
                  "summary": "Update source name only",
                  "value": {
                    "name": "Updated Source Name"
                  }
                },
                "updateBoth": {
                  "summary": "Update both name and config",
                  "value": {
                    "name": "Updated Source Name",
                    "config": {
                      "type": "WORDPRESS",
                      "required": {
                        "url": "https://updated-blog.example.com"
                      },
                      "masked": {
                        "username": "admin",
                        "application_password": "updated-application-password"
                      },
                      "sampling": {
                        "strategy": "RANDOM",
                        "limit": 25
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Source successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - validation failed"
          },
          "404": {
            "description": "Source not found"
          }
        },
        "summary": "Update a data source",
        "tags": [
          "Sources"
        ]
      },
      "delete": {
        "description": "Permanently delete a data source and all its associated data",
        "operationId": "SourcesController_deleteSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Source unique identifier",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Source successfully deleted"
          },
          "404": {
            "description": "Source not found"
          }
        },
        "summary": "Delete a data source",
        "tags": [
          "Sources"
        ]
      }
    },
...
    "/sources/{id}/runs": {
      "post": {
        "description": "Initiate a new data ingestion run for the specified source. This creates a new run ID and sets the runner status to PENDING.",
        "operationId": "SourcesController_startRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Source unique identifier",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          }
        },
        "summary": "Start a new ingestion run",
        "tags": [
          "Sources"
        ]
      }
    },
    "/sources/{id}/status": {
      "patch": {
        "description": "Compatibility wrapper that updates the current runner for a source. Only terminal statuses are allowed.",
        "operationId": "SourcesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Source unique identifier",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "COMPLETED",
                      "ERROR"
                    ],
                    "example": "COMPLETED"
                  }
                },
                "required": [
                  "status"
                ]
              },
              "examples": {
                "running": {
                  "summary": "Set to running",
                  "value": {
                    "status": "RUNNING"
                  }
                },
                "completed": {
                  "summary": "Mark as completed",
                  "value": {
                    "status": "COMPLETED"
                  }
                },
                "error": {
                  "summary": "Mark as error",
                  "value": {
                    "status": "ERROR"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          },
          "404": {
            "description": "Source not found"
          }
        },
        "summary": "Update runner status",
        "tags": [
          "Sources"
        ]
      }
    },
...
    "/assets/{id}": {
      "get": {
        "description": "Retrieve detailed information about a specific asset",
        "operationId": "AssetsController_getAsset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Asset unique identifier (deterministic UUID)",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetListItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "summary": "Get asset by ID",
        "tags": [
          "Assets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Classifyre API",
    "description": "Metadata ingestion and detection API for unstructured data sources. Supports WordPress, Slack, Sitemap, S3-Compatible Storage, Azure Blob Storage, Google Cloud Storage, PostgreSQL, MySQL, MSSQL, Oracle, Hive, Databricks, Snowflake, MongoDB, PowerBI, Tableau, Confluence, Jira, and Service Desk sources. Built-in detectors for secrets, PII, toxic content, NSFW images, broken links, and security threats.",
    "version": "1.0.0",
    "contact": {
      "name": "Classifyre Team",
      "url": "https://github.com/unstructured/classifyre",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateSourceDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "WORDPRESS",
              "SLACK",
              "SITEMAP",
              "S3_COMPATIBLE_STORAGE",
              "AZURE_BLOB_STORAGE",
              "GOOGLE_CLOUD_STORAGE",
              "POSTGRESQL",
              "MYSQL",
              "MSSQL",
              "ORACLE",
              "HIVE",
              "DATABRICKS",
              "SNOWFLAKE",
              "MONGODB",
              "POWERBI",
              "TABLEAU",
              "CONFLUENCE",
              "JIRA",
              "SERVICEDESK"
            ],
            "description": "The type of the source",
            "example": "WORDPRESS"
          },
          "name": {
            "type": "string",
            "description": "The name of the source",
            "example": "Production WordPress"
          },
          "config": {
            "type": "object",
            "description": "Configuration for the source (depends on type)",
            "example": {
              "type": "WORDPRESS",
              "required": {
                "url": "https://blog.example.com"
              },
              "masked": {
                "username": "admin",
                "application_password": "your-application-password"
              },
              "optional": {
                "content": {
                  "fetch_posts": true,
                  "fetch_pages": true
                }
              }
            }
          },
          "scheduleEnabled": {
            "type": "boolean",
            "description": "Whether to enable a recurring schedule for this source",
            "example": true
          },
          "scheduleCron": {
            "type": "string",
            "description": "5-field cron expression (required when scheduleEnabled is true)",
            "example": "30 1 * * *"
          },
          "scheduleTimezone": {
            "type": "string",
            "description": "IANA timezone for the cron schedule",
            "example": "UTC"
          }
        },
        "required": [
          "type",
          "name",
          "config"
        ]
      },
      "SourceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "example": "Production WordPress"
          },
          "type": {
            "type": "string",
            "example": "WORDPRESS"
          },
          "config": {
            "type": "object"
          },
          "currentRunnerId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "runnerStatus": {
            "type": "string",
            "example": "PENDING"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-31T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-31T10:00:00.000Z"
          },
          "scheduleEnabled": {
            "type": "boolean",
            "example": false
          },
          "scheduleCron": {
            "type": "string",
            "nullable": true,
            "example": "30 1 * * *"
          },
          "scheduleTimezone": {
            "type": "string",
            "nullable": true,
            "example": "UTC"
          },
          "scheduleNextAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "config",
          "currentRunnerId",
          "runnerStatus",
          "createdAt",
          "updatedAt",
          "scheduleEnabled",
          "scheduleCron",
          "scheduleTimezone",
          "scheduleNextAt"
        ]
      },
...
    }
  }
}